9.15修改

+++ miniprogram/pages/store_operation_update/store_operation_update.js
@@ -13,28 +13,45 @@ Page({
     price:"",
     detail:"",
     id:"",
-    belong:""
+    belong:"",
+    content:"",
+    img:[],
+    judge:""
   },
+
   delete_product:function(e){
     let that = this
-    db.collection('food_list').doc(that.options.id).remove({
-      success: function(res) {
-        console.log('删除成功',res.data)
-        wx.cloud.deleteFile({
-          fileList: that.data.img,
-          success: res => {
-            // handle success
-            console.log(res.fileList)
-          },
-          fail: err => {
-            // handle error
-          },
-        })
-        wx.redirectTo({
-          url: '../store_operation_gl/store_operation_gl',
-        })
+   
+    wx.showModal({
+      title: '提示',
+      content: '确定删除该商品吗？',
+      success: function (res) {
+        if (res.confirm) {//这里是点击了确定以
+          console.log('用户点击确定')
+          db.collection('food_list').doc(that.options.id).remove({
+            success: function(res) {
+              console.log('删除成功',res.data)
+              wx.cloud.deleteFile({
+                fileList: that.data.img,
+                success: res => {
+                  // handle success
+                  console.log(res.fileList)
+                },
+                fail: err => {
+                  // handle error
+                },
+              })
+              wx.redirectTo({
+                url: '../store_operation_gl/store_operation_gl',
+              })
+            }
+          })
+        } else {//这里是点击了取消以后
+          console.log('用户点击取消')
+        }
       }
     })
+    
   },
   // 上传图片
   upload_img:function(){
@@ -95,18 +112,22 @@ Page({
     let that = this
     let id = that.data.id
     console.log(e)
-    if(e.detail.value.name!==""&&e.detail.value.price!==""&&e.detail.value.fenlei!==""&&e.detail.value.detail!==""&&that.data.img.length!==0){
+    if(e.detail.value.name!==""&&e.detail.value.price!==""&&e.detail.value.detail!==""&&that.data.img.length!==0){
       db.collection('food_list').doc(this.options.id).update({
         data:{
           name:e.detail.value.name,
           price:e.detail.value.price,
           fenlei:e.detail.value.fenlei,
-          content:e.detail.value.detail,
+          content:e.detail.value.content,
           img:that.data.img,
+          product_contain:e.detail.value.detail
         },
         success:function(res){
           wx.showToast({
             title: '提交成功',
+          })
+          wx.redirectTo({
+            url: '../store_operation_gl/store_operation_gl',
           })
         },fail:function(res){
           console.log('提交失败',res),
@@ -126,9 +147,9 @@ Page({
    */
   onLoad: function (options) {
     let that = this
-    that.setData({
-      id:options.id
-    })
+    // that.setData({
+    //   id:options.id
+    // })
     db.collection('fenlei').get({
       success:function(res){
         console.log('分类获取成功',res)
@@ -145,9 +166,11 @@ Page({
         that.setData({
           name:res.data.name,
           price:res.data.price,
-          detail:res.data.content,
+          detail:res.data.product_contain,
           id:res.data._id,
-          belong:res.data.belong
+          belong:res.data.belong,
+          content:res.data.content,
+          img:res.data.img
         })
       },fail:function(res){
         console.log('分信息获取失败',res)
@@ -166,7 +189,35 @@ Page({
    * 生命周期函数--监听页面显示
    */
   onShow: function () {
-
+    let that = this
+    // that.setData({
+    //   id:id_1
+    // })
+    db.collection('fenlei').get({
+      success:function(res){
+        console.log('分类获取成功',res)
+        that.setData({
+          fenlei:res.data
+        })
+      },fail:function(res){
+        console.log('分类获取失败',res)
+      }
+    })
+    db.collection('food_list').doc(id_1).get({
+      success:function(res){
+        console.log('信息获取成功',res)
+        that.setData({
+          name:res.data.name,
+          price:res.data.price,
+          detail:res.data.product_contain,
+          id:res.data._id,
+          belong:res.data.belong,
+          content:res.data.content
+        })
+      },fail:function(res){
+        console.log('分信息获取失败',res)
+      }
+    })
   },
 
   /**
